moved image processor into own file
